House auth utils in auth path, error when resolved tenant/db don't match input
